Refactored execution_context_unit_test.ccp and moved redundant defined(CONFIG_BPF_JIT_DISABLED) blocks to a different file
