+++ libs/execution_context/ebpf_core_config.h
@@ -12,8 +12,7 @@ extern "C"
 {
 #endif
 
-// Shared variables between ebpf_core.c and ebpf_core_config.c
-// Assume enabled until we can query it.
+// Shared variable between ebpf_core.c and ebpf_core_config.c, defined in ebpf_core.c
 extern bool _ebpf_platform_hypervisor_code_integrity_enabled;
 
 #define PROTOCOL_NATIVE_MODE 1
