+++ tests/unit/test.vcxproj
@@ -214,6 +214,7 @@
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="..\..\libs\execution_context\unit\execution_context_unit_test.cpp" />
+    <ClCompile Include="..\..\libs\execution_context\unit\execution_context_unit_test_jit.cpp" />
     <ClCompile Include="..\..\libs\runtime\unit\platform_unit_test.cpp" />
     <ClCompile Include="..\..\libs\thunk\mock\mock.cpp" />
     <ClCompile Include="..\end_to_end\end_to_end.cpp" />
@@ -224,6 +225,7 @@
     <ClCompile Include="libbpf_test.cpp" />
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="..\..\libs\execution_context\unit\execution_context_unit_test_jit.h" />
     <ClInclude Include="..\..\libs\thunk\mock\mock.h" />
     <ClInclude Include="..\end_to_end\helpers.h" />
     <ClInclude Include="..\end_to_end\test_helper.hpp" />
